requirements.txt
aioimaplib>=1.1.0
voluptuous>=0.13.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
